+++ src/utils/__init__.py
@@ -12,6 +12,8 @@ from src.utils.trade_logger import (
 from src.utils.model_utils import download_model_if_missing, download_feature_list_if_missing
 from src.utils.env_utils import get_env_float
 from src.utils.gc_utils import maybe_collect
+from src.utils.data_utils import convert_thai_datetime, prepare_csv_auto
+from src.utils.resource_plan import get_resource_plan, save_resource_plan
 
 __all__ = [
     "get_session_tag",
@@ -24,6 +26,9 @@ __all__ = [
     "download_model_if_missing",
     "get_env_float",
     "download_feature_list_if_missing",
-    "get_env_float",
     "maybe_collect",
+    "convert_thai_datetime",
+    "prepare_csv_auto",
+    "get_resource_plan",
+    "save_resource_plan",
 ]